+++ tests/ut/ops/test_fused_moe.py
@@ -8,8 +8,10 @@ from torch import nn
 
 from vllm_ascend.ascend_forward_context import MoECommType
 from vllm_ascend.ops.fused_moe import fused_moe as fused_moe_module
-from vllm_ascend.ops.fused_moe.fused_moe import (
-    AscendMoERunner,
+from vllm_ascend.ops.fused_moe import routed_experts as routed_experts_module
+from vllm_ascend.ops.fused_moe.fused_moe import AscendMoERunner
+from vllm_ascend.ops.fused_moe.routed_experts import (
+    AscendRoutedExperts,
     AscendUnquantizedFusedMoEMethod,
     make_eplb_placement_config,
     use_multistage_eplb_load,
@@ -74,6 +76,34 @@ def test_ascend_unquantized_skips_upstream_modular_kernel_init():
     method = AscendUnquantizedFusedMoEMethod.__new__(AscendUnquantizedFusedMoEMethod)
 
     assert method.maybe_make_prepare_finalize() is None
+
+
+def test_ascend_routed_experts_owns_unquantized_method():
+    routed_experts = AscendRoutedExperts.__new__(AscendRoutedExperts)
+    routed_experts.tid2eid = object()
+    moe_config = MagicMock()
+
+    method = routed_experts._get_quant_method("model.layers.0.mlp", None, moe_config)
+
+    assert isinstance(method, AscendUnquantizedFusedMoEMethod)
+    assert method.moe is moe_config
+    assert method.tid2eid is routed_experts.tid2eid
+
+
+def test_ascend_routed_experts_passes_tid2eid_to_quant_config():
+    routed_experts = AscendRoutedExperts.__new__(AscendRoutedExperts)
+    routed_experts.tid2eid = object()
+    quant_config = MagicMock()
+    moe_config = MagicMock()
+
+    method = routed_experts._get_quant_method("model.layers.0.mlp", quant_config, moe_config)
+
+    assert method is quant_config.get_quant_method.return_value
+    quant_config.get_quant_method.assert_called_once_with(
+        routed_experts,
+        "model.layers.0.mlp",
+        tid2eid=routed_experts.tid2eid,
+    )
 
 
 def test_process_weights_after_loading_uses_version_specific_layout(
@@ -178,6 +208,90 @@ def test_runner_reduction_contract(monkeypatch, moe_comm_type, flash_comm_v1_ena
     assert runner._maybe_reduce_shared_expert_output(shared_output) is shared_output
 
 
+def test_routed_experts_no_shared_forward_impl_runs_current_flow(monkeypatch):
+    routed_experts = AscendRoutedExperts.__new__(AscendRoutedExperts)
+    hidden_states = torch.randn(2, 4)
+    prepared_hidden_states = torch.randn(2, 4)
+    router_logits = torch.randn(2, 3)
+    prepared_router_logits = torch.randn(2, 3)
+    routed_out = torch.randn(2, 4)
+    finalized = torch.randn(2, 4)
+    quant_method = MagicMock()
+    quant_method.quant_method = SimpleNamespace(quant_type=QuantType.NONE)
+    quant_method.apply.return_value = SimpleNamespace(
+        routed_out=routed_out,
+        expert_tokens=None,
+        group_list_type=1,
+        before_dispatch_evt=None,
+        before_gmm2_evt=None,
+        before_combine_evt=None,
+        swiglu_limit=0.0,
+    )
+    routed_experts.enable_npugraph_ex_static_kernel = False
+    routed_experts.enable_shared_expert_dp = False
+    routed_experts.quant_method = quant_method
+    routed_experts.top_k = 2
+    routed_experts.renormalize = True
+    routed_experts.use_grouped_topk = False
+    routed_experts.moe_config = SimpleNamespace(num_experts=4)
+    routed_experts.ascend_expert_map = None
+    routed_experts.topk_group = None
+    routed_experts.num_expert_group = None
+    routed_experts.custom_routing_function = None
+    routed_experts.scoring_func = "softmax"
+    routed_experts.routed_scaling_factor = 1.0
+    routed_experts.e_score_correction_bias = None
+    routed_experts.activation = "gelu"
+    routed_experts.apply_router_weight_on_input = True
+    routed_experts.log2phy = None
+    routed_experts.global_redundant_expert_num = 0
+    routed_experts.dynamic_eplb = False
+    moe_comm_method = MagicMock()
+    moe_comm_method.prepare.return_value = SimpleNamespace(
+        hidden_states=prepared_hidden_states,
+        router_logits=prepared_router_logits,
+        mc2_mask=None,
+        padded_hidden_states_shape=torch.Size([2, 4]),
+        pertoken_scale=None,
+    )
+    moe_comm_method.finalize.return_value = finalized
+    monkeypatch.setattr(
+        routed_experts_module,
+        "_EXTRA_CTX",
+        SimpleNamespace(
+            in_profile_run=False,
+            moe_comm_method=moe_comm_method,
+            flash_comm_v1_enabled=False,
+            eplb_heat_collection_status=False,
+        ),
+    )
+    monkeypatch.setattr(routed_experts_module, "get_forward_context", lambda: SimpleNamespace(all_moe_layers=None))
+
+    result = routed_experts.no_shared_forward_impl(
+        hidden_states=hidden_states,
+        router_logits=router_logits,
+    )
+
+    assert result is finalized
+    moe_comm_method.prepare.assert_called_once_with(
+        hidden_states=hidden_states,
+        router_logits=router_logits,
+        replace_allreduce=False,
+        enable_shared_expert_dp=False,
+        quant_type=QuantType.NONE,
+    )
+    quant_method.apply.assert_called_once()
+    assert quant_method.apply.call_args.kwargs["layer"] is routed_experts
+    assert quant_method.apply.call_args.kwargs["x"] is prepared_hidden_states
+    assert quant_method.apply.call_args.kwargs["router_logits"] is prepared_router_logits
+    assert quant_method.apply.call_args.kwargs["activation"] == "gelu"
+    moe_comm_method.finalize.assert_called_once_with(
+        hidden_states=routed_out,
+        reduce_results=False,
+        padded_hidden_states_shape=torch.Size([2, 4]),
+    )
+
+
 class _Projection(nn.Module):
     def forward(self, hidden_states):
         return hidden_states * 2.0 + 1.0, None
@@ -224,7 +338,7 @@ def test_shared_forward_impl_returns_current_runner_contract(monkeypatch, has_sh
         before_combine_evt=None,
         swiglu_limit=0.0,
     )
-    runner.no_shared_forward_impl = MagicMock(return_value=routed_result)
+    runner.routed_experts = SimpleNamespace(no_shared_forward_impl=MagicMock(return_value=routed_result))
     runner._forward_shared_experts = MagicMock(return_value=shared_out)
     current_stream = MagicMock()
 
@@ -233,9 +347,9 @@ def test_shared_forward_impl_returns_current_runner_contract(monkeypatch, has_sh
 
     result = runner.shared_forward_impl(hidden_states, router_logits)
 
-    runner.no_shared_forward_impl.assert_called_once_with(
-        hidden_states,
-        router_logits,
+    runner.routed_experts.no_shared_forward_impl.assert_called_once_with(
+        hidden_states=hidden_states,
+        router_logits=router_logits,
         return_with_event=True,
     )
     if has_shared_experts:
