+++ vllm_ascend/_310p/fused_moe/fused_moe.py
@@ -25,6 +25,7 @@ from vllm_ascend.ops.fused_moe.experts_selector import zero_experts_compute
 from vllm_ascend.ops.fused_moe.fused_moe import AscendMoERunner
 from vllm_ascend.ops.fused_moe.moe_comm_method import _MoECommMethods
 from vllm_ascend.ops.fused_moe.moe_runtime_args import build_fused_experts_input
+from vllm_ascend.ops.fused_moe.routed_experts import AscendRoutedExperts
 from vllm_ascend.quantization.quant_type import QuantType
 from vllm_ascend.utils import maybe_trans_nz
 
@@ -121,6 +122,13 @@ class AscendUnquantizedFusedMoEMethod310(UnquantizedFusedMoEMethod):
         return final_hidden_states
 
 
+class AscendRoutedExperts310(AscendRoutedExperts):
+    def _get_quant_method(self, prefix, quant_config, moe_config):
+        if quant_config is None:
+            return AscendUnquantizedFusedMoEMethod310(moe_config)
+        return quant_config.get_quant_method(self, prefix, tid2eid=self.tid2eid)
+
+
 class AscendMoERunner310(AscendMoERunner):
     def __init__(
         self,
@@ -153,10 +161,6 @@ class AscendMoERunner310(AscendMoERunner):
             tid2eid,
             n_shared_experts,
         )
-
-        if routed_experts.quant_config is None:
-            routed_experts.quant_method = AscendUnquantizedFusedMoEMethod310(self.moe_config)
-            self.quant_type = self._get_quant_type()
 
         self.multistream_overlap_shared_expert = False
         _MoECommMethods[MoECommType.ALLGATHER] = AllGatherCommImpl310(self.moe_config)
